Reorganized core modules in order to avoid nasty circular dependencies.
