+++ core/module.py
@@ -19,269 +19,15 @@ Copyright (c) the Qudi Developers. See the COPYRIGHT.txt file at the
 top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi/>
 """
 
-import copy
 import logging
 import warnings
 from fysom import Fysom  # provides a final state machine
 from collections import OrderedDict
-from enum import Enum
+
 from qtpy import QtCore
-from core.interfaces import InterfaceMethod
-
-
-class StatusVar:
-    """ This class defines a status variable that is loaded before activation
-        and saved after deactivation.
-    """
-
-    def __init__(self, name=None, default=None, *, var_name=None, constructor=None, representer=None):
-        """
-            @param name: identifier of the status variable when stored
-            @param default: default value for the status variable when a
-                saved version is not present
-            @param constructor: constructor function for variable, do loading type checks or conversion here
-            @param representer: representer function for status variable, do saving conversion here
-            @param var_name: name of the variable inside a running module. Only set this
-                if you know what you are doing!
-        """
-        self.var_name = var_name
-        if name is None:
-            self.name = var_name
-        else:
-            self.name = name
-
-        self.constructor_function = constructor
-        self.representer_function = representer
-        self.default = default
-
-    def copy(self, **kwargs):
-        """ Create a new instance of StatusVar with copied and updated values.
-
-            @param kwargs: Additional or overridden parameters for the constructor of this class
-        """
-        newargs = {'name': copy.copy(self.name), 'default': copy.copy(self.default),
-                   'constructor': self.constructor_function, 'representer': self.representer_function,
-                   'var_name': copy.copy(self.var_name)}
-        newargs.update(kwargs)
-        return StatusVar(**newargs)
-
-    def constructor(self, func):
-        """ This is the decorator for declaring constructor function for this StatusVar.
-
-            @param func: constructor function for this StatusVar
-            @return: return the original function so this can be used as a decorator
-        """
-        if callable(func):
-            self.constructor_function = func
-        return func
-
-    def representer(self, func):
-        """ This is the decorator for declaring a representer function for this StatusVar.
-
-            @param func: representer function for this StatusVar
-            @return: return the original function so this can be used as a decorator
-        """
-        if callable(func):
-            self.representer_function = func
-        return func
-
-
-class MissingOption(Enum):
-    """ Representation for missing ConfigOption """
-    error = -3
-    warn = -2
-    info = -1
-    nothing = 0
-
-
-class ConfigOption:
-    """ This class represents a configuration entry in the config file that is loaded before
-        module initalisation.
-    """
-
-    def __init__(self, name=None, default=None, *, var_name=None, missing='nothing',
-                 constructor=None, checker=None, converter=None):
-        """ Create a ConfigOption object.
-
-            @param name: identifier of the option in the configuration file
-            @param default: default value for the case that the option is not set
-                in the config file
-            @param var_name: name of the variable inside a running module. Only set this
-                if you know what you are doing!
-            @param missing: action to take when the option is not set. 'nothing' does nothing,
-                'warn' logs a warning, 'error' logs an error and prevents the module from loading
-            @param constructor: constructor function for complex config option behaviour
-            @param checker: static function that checks if value is ok
-            @param converter: static function that forces type interpretation
-        """
-        self.missing = MissingOption[missing]
-        self.var_name = var_name
-        if name is None:
-            self.name = var_name
-        else:
-            self.name = name
-
-        self.default = default
-        self.constructor_function = constructor
-        self.checker = checker
-        self.converter = converter
-
-    def copy(self, **kwargs):
-        """ Create a new instance of ConfigOption with copied values and update
-
-            @param kwargs: extra arguments or overrides for the constructor of this class
-        """
-        newargs = {'name': copy.copy(self.name), 'default': copy.copy(self.default),
-                   'var_name': copy.copy(self.var_name), 'missing': copy.copy(self.missing.name),
-                   'constructor': self.constructor_function, 'checker': self.checker, 'converter': self.converter}
-        newargs.update(kwargs)
-        return ConfigOption(**newargs)
-
-    def check(self, value):
-        """ If checker function set, check value. Else assume everything is ok.
-        """
-        if callable(self.checker):
-            return self.checker(value)
-        else:
-            return True
-
-    def convert(self, value):
-        """ If converter function set, convert value. Needs to raise exception on error.
-
-            @param value: value to convert (or not)
-
-            @return: converted value (or passthrough)
-        """
-        if callable(self.converter):
-            return self.converter(value)
-        else:
-            return value
-
-    def constructor(self, func):
-        """ This is the decorator for declaring a constructor function for this ConfigOption.
-
-            @param func: constructor function for this ConfigOption
-            @return: return the original function so this can be used as a decorator
-        """
-        if callable(func):
-            self.constructor_function = func
-        return func
-
-
-class Connector:
-    """ A connector where another module can be connected """
-
-    def __init__(self, *, name=None, interface=None, optional=False):
-        """
-            @param name: name of the connector
-            @param interface: interface class or name of the interface for this connector
-            @param (bool) optional: the optionality of the connector
-        """
-        self.name = name
-        self.interface = interface
-        self.obj = None
-        self.optional = optional
-
-    def __call__(self):
-        """ Return reference to the module that this connector is connected to. """
-        if self.obj is None:
-            if self.optional:
-                return None
-            raise Exception(
-                'Connector {0} (interface {1}) is not connected.'
-                ''.format(self.name, self.interface))
-
-        class ConnectedInterfaceProxy:
-            def __getattribute__(*args):
-                attr = getattr(self.obj, args[1])
-                if isinstance(attr, InterfaceMethod):
-                    return attr[self.interface]
-                else:
-                    return attr
-        return ConnectedInterfaceProxy()
-    
-    @property
-    def is_connected(self):
-        return self.obj is not None
-
-    def connect(self, target):
-        """ Check if target is connectable this connector and connect."""
-        if not isinstance(self.interface, str):
-            if isinstance(target, self.interface):
-                self.obj = target
-            else:
-                raise Exception(
-                    'Module {0} connected to connector {1} does not implement interface {2}.'
-                    ''.format(target, self.name, self.interface))
-        else:
-            self.obj = target
-
-    def disconnect(self):
-        """ Disconnect connector. """
-        self.obj = None
-
-    #def __repr__(self):
-    #    return '<{0}: name={1}, interface={2}, object={3}>'.format(self.__class__, self.name, self.ifname, self.obj)
-
-    def copy(self, **kwargs):
-        """ Create a new instance of Connector with copied values and update """
-        newargs = {'name': copy.copy(self.name), 'interface': copy.copy(self.interface),
-                   'optional': copy.copy(self.optional)}
-        newargs.update(kwargs)
-        return Connector(**newargs)
-
-
-class ModuleMeta(type(QtCore.QObject)):
-    """
-    Metaclass for Qudi modules
-    """
-
-    def __new__(cls, name, bases, attrs):
-        """
-        Collect declared Connectors, ConfigOptions and StatusVars into dictionaries.
-
-            @param mcs: class
-            @param name: name of class
-            @param bases: list of base classes of class
-            @param attrs: attributes of class
-
-            @return: new class with collected connectors
-        """
-
-        # collect meta info in dicts
-        connectors = OrderedDict()
-        config_options = OrderedDict()
-        status_vars = OrderedDict()
-
-        # Accumulate Connector, ConfigOption and StatusVar info from parent classes
-        for base in reversed(bases):
-            if hasattr(base, '_connectors'):
-                connectors.update(copy.deepcopy(base._connectors))
-            if hasattr(base, '_config_options'):
-                config_options.update(copy.deepcopy(base._config_options))
-            if hasattr(base, '_stat_var'):
-                status_vars.update(copy.deepcopy(base._stat_var))
-
-        # Collect this classes Connector and ConfigOption and StatusVar into dictionaries
-        for key, value in attrs.items():
-            if isinstance(value, Connector):
-                connectors[key] = value.copy(name=key)
-            elif isinstance(value, ConfigOption):
-                config_options[key] = value.copy(var_name=key)
-            elif isinstance(value, StatusVar):
-                status_vars[key] = value.copy(var_name=key)
-
-        attrs.update(connectors)
-        attrs.update(config_options)
-        attrs.update(status_vars)
-
-        # create a new class with the new dictionaries
-        new_class = super().__new__(cls, name, bases, attrs)
-        new_class._conn = connectors
-        new_class._config_options = config_options
-        new_class._stat_vars = status_vars
-
-        return new_class
+from .meta import ModuleMeta
+from .configoption import MissingOption
+from .connector import Connector
 
 
 class ModuleStateMachine(QtCore.QObject, Fysom):
