+++ core/qudikernel.py
@@ -31,7 +31,7 @@ import tempfile
 
 import config
 
-from parentpoller import ParentPollerUnix, ParentPollerWindows
+from .parentpoller import ParentPollerUnix, ParentPollerWindows
 
 rpyc.core.protocol.DEFAULT_CONFIG['allow_pickle'] = True
