+++ interface/first_test_interface.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 
 import abc
-from core.interfaces import InterfaceMetaclass, interface_method, abstract_interface_method
+from core.interface import interface_method, abstract_interface_method
+from core.meta import InterfaceMetaclass
 
 
 class FirstTestInterface(metaclass=InterfaceMetaclass):
