+++ interface/pulser_interface.py
@@ -22,7 +22,8 @@ top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi
 
 
 import abc
-from core.interfaces import InterfaceMetaclass, ScalarConstraint
+from core.meta import InterfaceMetaclass
+from core.interface import ScalarConstraint
 
 
 class PulserInterface(metaclass=InterfaceMetaclass):
